# apps/backend/docs/swagger.yaml
basePath: /
definitions:
  db.DevicesModel:
    properties:
      createdAt:
        type: string
      deviceType:
        type: string
      endpointArn:
        type: string
      fcmToken:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.NotficationStatus:
    enum:
    - Sent
    - Failed
    - Pending
    type: string
    x-enum-varnames:
    - NotficationStatusSent
    - NotficationStatusFailed
    - NotficationStatusPending
  db.NotificationModel:
    properties:
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      metadata:
        items:
          type: integer
        type: array
      status:
        $ref: '#/definitions/db.NotficationStatus'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
      website:
        $ref: '#/definitions/db.WebsiteModel'
      websiteId:
        type: string
    type: object
  db.OrganizationMemberModel:
    properties:
      createdAt:
        type: string
      id:
        type: string
      organization:
        $ref: '#/definitions/db.OrganizationModel'
      organizationId:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.OrganizationModel:
    properties:
      admin:
        $ref: '#/definitions/db.UserModel'
      adminId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/db.OrganizationMemberModel'
        type: array
      name:
        type: string
      updatedAt:
        type: string
      websites:
        items:
          $ref: '#/definitions/db.WebsiteModel'
        type: array
    type: object
  db.RegionModel:
    properties:
      createdAt:
        type: string
      regionId:
        type: string
      regionName:
        type: string
      ticks:
        items:
          $ref: '#/definitions/db.WebsiteTicksModel'
        type: array
      updatedAt:
        type: string
    type: object
  db.SubscriptionModel:
    properties:
      createdAt:
        type: string
      deviceType:
        type: string
      id:
        type: string
      topicArn:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.UserModel:
    properties:
      createdAt:
        type: string
      devices:
        items:
          $ref: '#/definitions/db.DevicesModel'
        type: array
      email:
        type: string
      id:
        type: string
      notifications:
        items:
          $ref: '#/definitions/db.NotificationModel'
        type: array
      organizationMemberships:
        items:
          $ref: '#/definitions/db.OrganizationMemberModel'
        type: array
      organizationsAsAdmin:
        items:
          $ref: '#/definitions/db.OrganizationModel'
        type: array
      password:
        type: string
      subscriptions:
        items:
          $ref: '#/definitions/db.SubscriptionModel'
        type: array
      updatedAt:
        type: string
      userProfile:
        $ref: '#/definitions/db.UserProfileModel'
    type: object
  db.UserProfileModel:
    properties:
      avatar:
        type: string
      bio:
        type: string
      createdAt:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      phone:
        type: string
      preferences:
        items:
          type: integer
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/db.UserModel'
      userId:
        type: string
    type: object
  db.WebsiteModel:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isMonitoringEnabled:
        type: boolean
      notifications:
        items:
          $ref: '#/definitions/db.NotificationModel'
        type: array
      organization:
        $ref: '#/definitions/db.OrganizationModel'
      organizationId:
        type: string
      status:
        $ref: '#/definitions/db.WebsiteStatus'
      ticks:
        items:
          $ref: '#/definitions/db.WebsiteTicksModel'
        type: array
      updatedAt:
        type: string
      url:
        type: string
      websiteName:
        type: string
    type: object
  db.WebsiteStatus:
    enum:
    - Up
    - Down
    - Unknown
    type: string
    x-enum-varnames:
    - WebsiteStatusUp
    - WebsiteStatusDown
    - WebsiteStatusUnknown
  db.WebsiteTicksModel:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      latency:
        type: integer
      region:
        $ref: '#/definitions/db.RegionModel'
      upStatus:
        $ref: '#/definitions/db.WebsiteStatus'
      updatedAt:
        type: string
      website:
        $ref: '#/definitions/db.WebsiteModel'
      websiteId:
        type: string
      websiteRegionId:
        type: string
    type: object
...
  internal_handlers.AssignWebsiteRegionsRequest:
    properties:
      regionIds:
        items:
          type: string
        type: array
    type: object
  internal_handlers.AssignWebsiteRegionsResponse:
    properties:
      regionIds:
        items:
          type: string
        type: array
      websiteId:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  title: My API Name
  version: "1.0"
paths:
...
  /api/incidents:
    get:
      description: Returns global incident history with optional filters
      parameters:
      - description: Filter by website ID
        in: query
        name: websiteId
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: string
      - description: down,unknown,degraded,up
        in: query
        name: status
        type: string
      - description: YYYY-MM-DD
        in: query
        name: startDate
        type: string
      - description: YYYY-MM-DD
        in: query
        name: endDate
        type: string
      - description: Offset
        in: query
        name: skip
        type: integer
      - description: Limit
        in: query
        name: take
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List incidents
      tags:
      - incidents
  /api/monitoring/{id}:
    post:
      consumes:
      - application/json
      description: Enables or disables monitoring for a given website
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.WebsiteModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Toggle website monitoring
      tags:
      - monitoring
...
  /api/websites/{id}/details:
    get:
      description: Retrieves uptime ticks and latency data for a website within a
        date range
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      - description: Number of records to skip
        in: query
        name: skip
        type: integer
      - description: Number of records to take
        in: query
        name: take
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.WebsiteTicksModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get detailed website metrics
      tags:
      - websites
  /api/websites/{id}/regions:
    put:
      consumes:
      - application/json
      description: Replaces the region assignment used by monitoring workers for the
        provided website
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      - description: Region IDs to assign
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.AssignWebsiteRegionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AssignWebsiteRegionsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Assign monitoring regions for a website
      tags:
      - websites
swagger: "2.0"